pyproject.toml
[project]
name = "ydl-podcast"
version = "1.12.4"
description = "A simple tool to generate Podcast-like RSS feeds from youtube (or other youtube-dl supported services) channels, using youtube-dl/yt-dlp"
license = { text = "MIT" }
authors = [
    { name = "nbr23", email = "max@23.tf" }
]
readme = "README.md"
requires-python = ">=3.9"
keywords = ["podcast", "youtube-dl"]
dependencies = [
    "yt-dlp>=2026.03.17",
    "PyYAML==6.0.2",
    "Jinja2>=3.1.2",
    "Pillow==12.1.1",
]

[project.optional-dependencies]
yt-dlp = ["yt-dlp>=2026.03.17"]
youtube-dl = ["youtube-dl"]

[project.urls]
homepage = "https://github.com/nbr23/ydl-podcast"
repository = "https://github.com/nbr23/ydl-podcast"

[project.scripts]
ydl-podcast = "ydl_podcast.__main__:main"

[dependency-groups]
test = ["pytest>=8.0"]

[tool.uv]
environments = ["python_version >= '3.12'"]

[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"
